[Presets] Separate Presets from policies
Moves the `PresetDefaults` base class out from
`sos/policies/__init__.py` and into a new `sos/presets` subdir. Further,
removes the presets defined in the RH-family policies into a new
`sos/presets/redhat.py` subdir to allow for ease of maintenance and
tracking.

Related: #2349
